Refactor database.py into a module
